requirements.txt
cassandra-driver==3.25.0
pydantic-1.9.0